+++ patchable-macro/src/lib.rs
@@ -17,7 +17,13 @@ pub fn derive_patchable(input: TokenStream) -> TokenStream {
     derive_with(input, |ctx| {
         let patch_struct_def = ctx.build_patch_struct();
         let patchable_trait_impl = ctx.build_patchable_trait_impl();
-        let from_struct_impl = ctx.build_from_trait_impl();
+        let from_struct_impl = cfg!(feature = "impl_from").then(|| {
+            let from_struct_impl = ctx.build_from_trait_impl();
+            quote! {
+                #[automatically_derived]
+                #from_struct_impl
+            }
+        });
 
         quote! {
             const _: () = {
@@ -25,7 +31,6 @@ pub fn derive_patchable(input: TokenStream) -> TokenStream {
                 #patch_struct_def
                 #[automatically_derived]
                 #patchable_trait_impl
-                #[automatically_derived]
                 #from_struct_impl
             };
         }