+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- `impl_from` feature to enable deriving `From<Struct>` for `StructPatch`
+
 ## [0.5.0] - 2026-01-27
 
 ### Changed